28-07-2021

+++ src/main/java/com/asl/controller/PurchaseReturnController.java
@@ -2,7 +2,6 @@ package com.asl.controller;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -26,7 +25,6 @@ import com.asl.entity.PogrnHeader;
 import com.asl.enums.CodeType;
 import com.asl.enums.ResponseStatus;
 import com.asl.enums.TransactionCodeType;
-import com.asl.model.ServiceException;
 import com.asl.service.PocrnService;
 import com.asl.service.PogrnService;
 import com.asl.service.XcodesService;
@@ -121,11 +119,29 @@ public class PurchaseReturnController extends ASLAbstractController {
 		return pocrnService.save(responseHelper, pocrnHeader, pgh);
 	}
 
-	@GetMapping("/pocrndetail/{xcrnnum}")
-	public String reloadPocrnDetailTable(@PathVariable String xcrnnum, Model model) {
-		model.addAttribute("pocrnDetailsList", pocrnService.findPocrnDetailByXcrnnum(xcrnnum));
-		model.addAttribute("pocrnheader", pocrnService.findPocrnHeaderByXcrnnum(xcrnnum));
-		return "pages/procurement/grnreturn/pocrn::pocrndetailtable";
+	@PostMapping("/archive/{xcrnnum}")
+	public  @ResponseBody Map<String, Object> archive(@PathVariable String xcrnnum) {
+		Pocrnheader pch = pocrnService.findPocrnHeaderByXcrnnum(xcrnnum);
+		if(pch == null) {
+			responseHelper.setErrorStatusAndMessage("Can't find Purchase return");
+			return responseHelper.getResponse();
+		}
+
+		List<Pocrndetail> crndetails = pocrnService.findPocrnDetailByXcrnnum(xcrnnum);
+		if(crndetails != null && !crndetails.isEmpty()) {
+			responseHelper.setErrorStatusAndMessage("Delete purchase return details first");
+			return responseHelper.getResponse();
+		}
+
+		long hcount = pocrnService.deletePocrnHeader(xcrnnum);
+		if(hcount == 0) {
+			responseHelper.setErrorStatusAndMessage("Can't Delete Purchase Return");
+			return responseHelper.getResponse();
+		}
+
+		responseHelper.setSuccessStatusAndMessage("Purchase Return Deleted Successfully");
+		responseHelper.setRedirectUrl("/procurements/purchasereturn");
+		return responseHelper.getResponse();
 	}
 
 	@GetMapping("{xcrnnum}/pocrndetail/{xrow}/show")
@@ -141,12 +157,7 @@ public class PurchaseReturnController extends ASLAbstractController {
 			model.addAttribute("pocrndetail", pocrndetail);
 		} else {
 			Pocrndetail pocrndetail = pocrnService.findPocrnDetailByXcrnnumAndXrow(xcrnnum, Integer.parseInt(xrow));
-			if (pocrndetail == null) {
-				pocrndetail = new Pocrndetail();
-				pocrndetail.setXcrnnum(xcrnnum);
-				pocrndetail.setXqtyord(BigDecimal.ONE.setScale(2, RoundingMode.DOWN));
-				pocrndetail.setXrate(BigDecimal.ZERO.setScale(2, RoundingMode.DOWN));
-			}
+			pocrndetail.setPrevqty(pocrndetail.getXqtyord() == null ? BigDecimal.ZERO : pocrndetail.getXqtyord());
 			model.addAttribute("pocrndetail", pocrndetail);
 		}
 
@@ -160,34 +171,85 @@ public class PurchaseReturnController extends ASLAbstractController {
 			return responseHelper.getResponse();
 		}
 
-		// modify line amount
-		pocrnDetail.setXlineamt(pocrnDetail.getXqtygrn().multiply(pocrnDetail.getXrate().setScale(2, RoundingMode.DOWN)));
+		if(StringUtils.isBlank(pocrnDetail.getXitem())) {
+			responseHelper.setErrorStatusAndMessage("Item selection required");
+			return responseHelper.getResponse();
+		}
+
+		if(pocrnDetail.getXqtyord() == null || BigDecimal.ZERO.equals(pocrnDetail.getXqtyord()) || pocrnDetail.getXqtyord().compareTo(BigDecimal.ZERO) == -1){
+			responseHelper.setErrorStatusAndMessage("Return quantity should be greater then 0");
+			return responseHelper.getResponse();
+		}
+
+		// validate return qty
+		Pocrnheader pch = pocrnService.findPocrnHeaderByXcrnnum(pocrnDetail.getXcrnnum());
+		if(pch == null) {
+			responseHelper.setErrorStatusAndMessage("Can't find Purchase return");
+			return responseHelper.getResponse();
+		}
+		PogrnDetail pogrndetail = pogrnService.findPogrnDetailByXgrnnumAndXrow(pch.getXgrnnum(), pocrnDetail.getXdocrow());
+		if(pogrndetail == null) {
+			responseHelper.setErrorStatusAndMessage("Can't find Puchase order detail for this item row");
+			return responseHelper.getResponse();
+		}
+		// calculate and update poordetail xqtygrn 
+		if(pocrnDetail.getXrow() != 0) {
+			BigDecimal diff1 = pocrnDetail.getPrevqty().subtract(pocrnDetail.getXqtyord());
+			BigDecimal diff2 = pogrndetail.getXqtyprn().subtract(diff1);
+			if(diff2.compareTo(pogrndetail.getXqtygrn()) == 1) {
+				responseHelper.setErrorStatusAndMessage("Return quantity can't be greater then GRN quantity");
+				return responseHelper.getResponse();
+			}
+			pogrndetail.setXqtyprn(diff2);
+		} else {
+			pogrndetail.setXqtyprn(pogrndetail.getXqtyprn().add(pocrnDetail.getXqtyord()));
+			if(pogrndetail.getXqtyprn().compareTo(pogrndetail.getXqtygrn()) == 1) {
+				responseHelper.setErrorStatusAndMessage("Purchase return quantity can't be greater then GRN quantity");
+				return responseHelper.getResponse();
+			}
+		}
+
+		long count2 = pogrnService.updateDetail(pogrndetail);
+		if(count2 == 0) {
+			responseHelper.setErrorStatusAndMessage("Can't update purchase return qty to GRN detail");
+			return responseHelper.getResponse();
+		}
+
+		pocrnDetail.setXitem(pocrnDetail.getXitem().split("\\|")[0]);
+		pocrnDetail.setXlineamt(pocrnDetail.getXqtyord().multiply(pocrnDetail.getXrate().setScale(2, RoundingMode.DOWN)));
+
 
 		// if existing
-		Pocrndetail existDetail = pocrnService.findPocrnDetailByXcrnnumAndXrow(pocrnDetail.getXcrnnum(),
-				pocrnDetail.getXrow());
+		Pocrndetail existDetail = pocrnService.findPocrnDetailByXcrnnumAndXrow(pocrnDetail.getXcrnnum(),pocrnDetail.getXrow());
 		if (existDetail != null) {
 			BeanUtils.copyProperties(pocrnDetail, existDetail, "xcrnnum", "xrow");
 			long count = pocrnService.updateDetail(existDetail);
 			if (count == 0) {
-				responseHelper.setStatus(ResponseStatus.ERROR);
+				responseHelper.setErrorStatusAndMessage("Can't update return item");
 				return responseHelper.getResponse();
 			}
-			responseHelper.setRedirectUrl("/procurements/purchasereturn/" + pocrnDetail.getXcrnnum());
-			responseHelper.setSuccessStatusAndMessage("PRN item updated successfully");
+
+			responseHelper.setReloadSectionIdWithUrl("pocrndetailtable", "/procurements/purchasereturn/pocrndetail/" + pocrnDetail.getXcrnnum());
+			responseHelper.setSuccessStatusAndMessage("Return item updated successfully");
 			return responseHelper.getResponse();
 		}
 
 		// if new detail
 		long count = pocrnService.saveDetail(pocrnDetail);
 		if (count == 0) {
-			responseHelper.setStatus(ResponseStatus.ERROR);
+			responseHelper.setErrorStatusAndMessage("Can't save return item");
 			return responseHelper.getResponse();
 		}
-		responseHelper.setRedirectUrl("/procurements/purchasereturn/" + pocrnDetail.getXcrnnum());
-		responseHelper.setSuccessStatusAndMessage("PRN item saved successfully");
-
+		responseHelper.setReloadSectionIdWithUrl("pocrndetailtable", "/procurements/purchasereturn/pocrndetail/" + pocrnDetail.getXcrnnum());
+		responseHelper.setSuccessStatusAndMessage("Return item saved successfully");
 		return responseHelper.getResponse();
+	}
+
+	@GetMapping("/pocrndetail/{xcrnnum}")
+	public String reloadPocrnDetailTable(@PathVariable String xcrnnum, Model model) {
+		model.addAttribute("pocrnDetailsList", pocrnService.findPocrnDetailByXcrnnum(xcrnnum));
+		model.addAttribute("pocrnheader", pocrnService.findPocrnHeaderByXcrnnum(xcrnnum));
+		return "pages/procurement/grnreturn/pocrn::pocrndetailtable";
 	}
 
 	@PostMapping("{xcrnnum}/pocrndetail/{xrow}/delete")
@@ -198,6 +260,23 @@ public class PurchaseReturnController extends ASLAbstractController {
 			return responseHelper.getResponse();
 		}
 
+		Pocrnheader pch = pocrnService.findPocrnHeaderByXcrnnum(pd.getXcrnnum());
+		if(pch == null) {
+			responseHelper.setErrorStatusAndMessage("Can't find Purchase return");
+			return responseHelper.getResponse();
+		}
+		PogrnDetail pogrndetail = pogrnService.findPogrnDetailByXgrnnumAndXrow(pch.getXgrnnum(), pd.getXdocrow());
+		if(pogrndetail == null) {
+			responseHelper.setErrorStatusAndMessage("Can't find Puchase order detail for this item row");
+			return responseHelper.getResponse();
+		}
+		pogrndetail.setXqtyprn(pogrndetail.getXqtyprn().subtract(pd.getXqtyord()));
+		long count2 = pogrnService.updateDetail(pogrndetail);
+		if(count2 == 0) {
+			responseHelper.setErrorStatusAndMessage("Can't update Return qty to GRN detail");
+			return responseHelper.getResponse();
+		}
+
 		long count = pocrnService.deleteDetail(pd);
 		if (count == 0) {
 			responseHelper.setStatus(ResponseStatus.ERROR);
@@ -205,35 +284,37 @@ public class PurchaseReturnController extends ASLAbstractController {
 		}
 
 		responseHelper.setSuccessStatusAndMessage("Deleted successfully");
-		responseHelper.setRedirectUrl("/procurements/purchasereturn/" + xcrnnum);
+		responseHelper.setReloadSectionIdWithUrl("pocrndetailtable", "/procurements/purchasereturn/pocrndetail/" + pd.getXcrnnum());
 		return responseHelper.getResponse();
 	}
 
-	@GetMapping("/confirmprn/{xcrnnum}")
+	@PostMapping("/confirmcrn/{xcrnnum}")
 	public @ResponseBody Map<String, Object> confirmgrn(@PathVariable String xcrnnum) {
-		if (StringUtils.isBlank(xcrnnum)) {
-			responseHelper.setStatus(ResponseStatus.ERROR);
-			return responseHelper.getResponse();
-		}
-		// Validate
-
-		// Get PocrnHeader record by Xcrnnum
 		Pocrnheader pocrnHeader = pocrnService.findPocrnHeaderByXcrnnum(xcrnnum);
-		PogrnHeader pogrnHeader = pogrnService.findPogrnHeaderByXgrnnum(pocrnHeader.getXgrnnum());
-		List<Pocrndetail> pocrnHeaderList = pocrnService.findPocrnDetailByXcrnnum(xcrnnum);
+		if(pocrnHeader == null) {
+			responseHelper.setErrorStatusAndMessage("GRN not found in this system");
+			return responseHelper.getResponse();
+		}
 
-		if (!"Open".equalsIgnoreCase(pocrnHeader.getXstatusgrn())) {
-			responseHelper.setErrorStatusAndMessage("PRN already confirmed");
+		// validate
+		if (StringUtils.isBlank(pocrnHeader.getXcus())) {
+			responseHelper.setErrorStatusAndMessage("Supplier required");
 			return responseHelper.getResponse();
 		}
-		if (pocrnHeaderList.size() == 0) {
-			responseHelper.setErrorStatusAndMessage("Please add details!");
+		if ("Confirmed".equalsIgnoreCase(pocrnHeader.getXstatuscrn())) {
+			responseHelper.setErrorStatusAndMessage("Purchase Return already confirmed");
 			return responseHelper.getResponse();
 		}
+
+		List<Pocrndetail> pocrndetailsList = pocrnService.findPocrnDetailByXcrnnum(xcrnnum);
+		if(pocrndetailsList == null || pocrndetailsList.isEmpty()) {
+			responseHelper.setErrorStatusAndMessage("Purchase Return has no item details");
+			return responseHelper.getResponse();
+		}
+
 		String p_seq;
 		if (!"Confirmed".equalsIgnoreCase(pocrnHeader.getXstatuscrn())) {
-			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(),
-					TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
+			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(), TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
 			pocrnService.procConfirmCRN(xcrnnum, p_seq);
 			// Error check for procConfirmCRN
 			String em = getProcedureErrorMessages(p_seq);
@@ -242,9 +323,8 @@ public class PurchaseReturnController extends ASLAbstractController {
 				return responseHelper.getResponse();
 			}
 
-			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(),
-					TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
-			pocrnService.procIssuePricing(pogrnHeader.getXdocnum().toString(), pocrnHeader.getXwh(), p_seq);
+			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(), TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
+			pocrnService.procIssuePricing(pocrnHeader.getXgrnnum(), pocrnHeader.getXwh(), p_seq);
 			// Error check for procIssuePricing
 			em = getProcedureErrorMessages(p_seq);
 			if (StringUtils.isNotBlank(em)) {
@@ -254,8 +334,7 @@ public class PurchaseReturnController extends ASLAbstractController {
 
 		}
 		if (!"Confirmed".equalsIgnoreCase(pocrnHeader.getXstatusap())) {
-			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(),
-					TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
+			p_seq = xtrnService.generateAndGetXtrnNumber(TransactionCodeType.PROC_ERROR.getCode(), TransactionCodeType.PROC_ERROR.getdefaultCode(), 6);
 			pocrnService.procTransferPRtoAP(xcrnnum, p_seq);
 			// Error check for procTransferPRtoAP
 			String em = getProcedureErrorMessages(p_seq);
@@ -265,7 +344,7 @@ public class PurchaseReturnController extends ASLAbstractController {
 			}
 		}
 
-		responseHelper.setSuccessStatusAndMessage("PRN Confirmed successfully");
+		responseHelper.setSuccessStatusAndMessage("Purchase return confirmed successfully");
 		responseHelper.setRedirectUrl("/procurements/purchasereturn/" + xcrnnum);
 		return responseHelper.getResponse();
 	}
